Fix form method
The form for resending email confirmation currently works by accedent. The form has a option method: :post, which makes it to use post method. However, there's an inner form_for tag, which adds a hidden field _method to the form and it contains value "put", which makes the form actually to use put method. This will cause us issues with Rails 4, since it changes the default "put" method to "patch"

Fix:

- Remove the inner "form_form" and replace it with "fields_for".
- Change method: :post to method: :put