Add hidden class to date_select instead of discard_day, discard_year, etc.
According to date_select documentation:

:discard_day - Set to true if you don't want to show a day select. This includes the day as a hidden field instead of showing a select field. Also note that this implicitly sets the day to be the first of the given month in order to not create invalid dates like 31 February.

That's why we can not use discard_day as a hidden field to pass the date data with the payment information. Instead, we can add a hidden class to the date selector.

This behaviour (defaulting to first day of the month) is probably added to Rails 4.