Unique community_memberships.person_id
Person and CommunityMemberhips have 1-to-1 relationship. Thus, the `person_id`in `community_memberships` should be unique globally, not only per community. Also, we want to be as little dependant on `community_id` in the `community_memberships` table, because in the end that's duplicated information which can be found in `people` table and we want to get rid of that duplication.